backend/requirements.txt
django-cors-headers==3.5.0
djangorestframework==3.12.2
Pillow==8.0.1
pipdeptree==2.0.0
setuptools==41.2.0
dj-database-url==0.5.0
django-cors-headers==3.5.0
djangorestframework==3.12.2
gunicorn==20.1.0
Pillow==8.0.1
pipdeptree==2.0.0
whitenoise==5.2.0
